atualizando logica de anunciosloc

+++ anunciosloc_server/src/main/java/com/anunciosloc/anunciosloc_server/dto/CoordenadaResponse.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class CoordenadaResponse {
     private Double latitude;
     private Double longitude;
-    private Integer raio;
+    private Double raio;
     private String ssidWifi;
 }
